+++ tests/test_asgi.py
@@ -114,6 +114,8 @@ def test_out_of_range_error(application, test_files):
     asyncio.run(application(scope, receive, send))
     assert send.status == 416
     assert send.headers[b"content-range"] == b"bytes */%d" % len(test_files.js_content)
+    assert send.headers[b"access-control-allow-origin"] == b"*"
+    assert b"cache-control" in send.headers
 
 
 def test_wrong_method_type(application, test_files):