+++ src/servestatic/__init__.py
@@ -3,6 +3,6 @@ from __future__ import annotations
 from servestatic.asgi import ServeStaticASGI
 from servestatic.wsgi import ServeStatic
 
-__version__ = "4.3.2"
+__version__ = "4.3.3"
 
 __all__ = ["ServeStatic", "ServeStaticASGI"]